bazel and AbstractTriggerTest cleanup
