+++ triggers/tests/BUILD.bazel
@@ -20,7 +20,6 @@ da_scala_library(
     srcs = [
         "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractFuncTests.scala",
         "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractTriggerTest.scala",
-        "src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractTriggerTestWithCanton.scala",
     ],
     scala_deps = [
         "@maven//:com_typesafe_akka_akka_stream",
@@ -33,26 +32,19 @@ da_scala_library(
     ],
     deps = [
         "//bazel_tools/runfiles:scala_runfiles",
-        "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
         "//daml-lf/integration-test-lib",
         "//daml-lf/interpreter",
-        "//daml-lf/language",
         "//daml-lf/transaction",
         "//language-support/scala/bindings",
         "//language-support/scala/bindings-akka",
         "//ledger-api/rs-grpc-bridge",
         "//ledger-api/testing-utils",
-        "//ledger/ledger-api-auth",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-domain",
-        "//ledger/ledger-runner-common",
         "//ledger/participant-integration-api",
-        "//ledger/sandbox-on-x",
-        "//ledger/sandbox-on-x:sandbox-on-x-test-lib",
         "//libs-scala/ledger-resources",
         "//libs-scala/ports",
-        "//test-common",
         "//triggers/runner:trigger-runner-lib",
         "@maven//:org_scalatest_scalatest_compatible",
     ],
@@ -165,8 +157,6 @@ EOF
                 "//ledger/ledger-configuration",
                 "//ledger/ledger-runner-common",
                 "//ledger/participant-integration-api",
-                "//ledger/sandbox-on-x",
-                "//ledger/sandbox-on-x:sandbox-on-x-test-lib",
                 "//libs-scala/contextualized-logging",
                 "//libs-scala/ledger-resources",
                 "//libs-scala/logging-entries",
@@ -321,8 +311,6 @@ da_scala_test_suite(
         "//ledger/ledger-configuration",
         "//ledger/ledger-runner-common",
         "//ledger/participant-integration-api",
-        "//ledger/sandbox-on-x",
-        "//ledger/sandbox-on-x:sandbox-on-x-test-lib",
         "//libs-scala/contextualized-logging",
         "//libs-scala/ledger-resources",
         "//libs-scala/logging-entries",
@@ -375,8 +363,6 @@ da_scala_test_suite(
         "//ledger/ledger-configuration",
         "//ledger/ledger-runner-common",
         "//ledger/participant-integration-api",
-        "//ledger/sandbox-on-x",
-        "//ledger/sandbox-on-x:sandbox-on-x-test-lib",
         "//libs-scala/contextualized-logging",
         "//libs-scala/ledger-resources",
         "//libs-scala/logging-entries",
