+++ triggers/tests/src/test/scala/com/digitalasset/daml/lf/engine/trigger/simulation/CatAndFoodTriggerSimulation.scala
@@ -11,6 +11,7 @@ import com.daml.ledger.api.refinements.ApiTypes.Party
 import com.daml.lf.engine.trigger.simulation.TriggerMultiProcessSimulation.TriggerSimulationConfig
 import com.daml.lf.engine.trigger.simulation.process.ledger.{LedgerExternalAction, LedgerProcess}
 import com.daml.lf.engine.trigger.simulation.process.TriggerProcessFactory
+import com.daml.lf.engine.trigger.test.AbstractTriggerTest
 import com.daml.lf.speedy.SValue
 import org.scalacheck.Gen
 import scalaz.syntax.tag._
@@ -22,6 +23,7 @@ class CatAndFoodTriggerSimulation
     extends TriggerMultiProcessSimulation
     with CatTriggerResourceUsageTestGenerators {
 
+  import AbstractTriggerTest._
   import CatAndFoodTriggerSimulation._
 
   // For demonstration purposes, we only run the simulation for 30 seconds
