+++ triggers/tests/src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractTriggerTest.scala
@@ -7,7 +7,9 @@ package trigger
 package test
 
 import akka.stream.Materializer
-import akka.stream.scaladsl.Sink
+
+import java.util.UUID
+import akka.stream.scaladsl.{Sink, Source}
 import com.daml.bazeltools.BazelRunfiles
 import com.daml.ledger.api.refinements.ApiTypes.{ApplicationId, Party}
 import com.daml.ledger.api.v1.command_service.SubmitAndWaitRequest
@@ -22,30 +24,34 @@ import com.daml.ledger.client.configuration.{
   LedgerClientConfiguration,
   LedgerIdRequirement,
 }
-import com.daml.ledger.sandbox.SandboxOnXForTest.ParticipantId
-import com.daml.lf.archive.DarDecoder
 import com.daml.lf.data.Ref._
 import com.daml.lf.engine.trigger.TriggerRunnerConfig.DefaultTriggerRunnerConfig
+import com.daml.lf.integrationtest.CantonFixture
 import com.daml.lf.speedy.SValue
 import com.daml.lf.speedy.SValue._
-import com.daml.platform.sandbox.{SandboxBackend, SandboxRequiringAuthorizationFuns}
-import com.daml.platform.sandbox.services.TestCommands
 import org.scalatest._
 import scalaz.syntax.tag._
-import com.daml.platform.sandbox.fixture.SandboxFixture
 
-import java.util.UUID
+import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 
-trait AbstractTriggerTest
-    extends SandboxFixture
-    with SandboxBackend.Postgresql
-    with TestCommands
-    with SandboxRequiringAuthorizationFuns {
+trait AbstractTriggerTest extends CantonFixture {
   self: Suite =>
 
-  protected implicit val applicationId: ApplicationId = RunnerConfig.DefaultApplicationId
+  import CantonFixture._
+
+  private[this] lazy val darFile =
+    Try(BazelRunfiles.requiredResource("triggers/tests/acs.dar").toPath) match {
+      case Success(value) => Right(value)
+      case Failure(_) => Left(BazelRunfiles.requiredResource("triggers/tests/acs-1.dev.dar").toPath)
+    }
+
+  final override protected lazy val darFiles = List(darFile.merge)
+  final override protected lazy val devMode = darFile.isLeft
+
+  implicit override protected lazy val applicationId: ApplicationId =
+    RunnerConfig.DefaultApplicationId
 
   protected def toHighLevelResult(s: SValue) = s match {
     case SRecord(_, _, values) if values.size == 6 =>
@@ -60,57 +66,36 @@ trait AbstractTriggerTest
     case _ => throw new IllegalArgumentException(s"Expected record with 6 fields but got $s")
   }
 
-  protected def ledgerClientConfiguration =
+  protected def ledgerClientConfiguration: LedgerClientConfiguration =
     LedgerClientConfiguration(
-      applicationId = ApplicationId.unwrap(applicationId),
+      applicationId = applicationId.unwrap,
       ledgerIdRequirement = LedgerIdRequirement.none,
       commandClient = CommandClientConfiguration.default,
       token = None,
     )
 
-  protected def ledgerClientChannelConfiguration =
+  protected def ledgerClientChannelConfiguration: LedgerClientChannelConfiguration =
     LedgerClientChannelConfiguration.InsecureDefaults
 
-  protected def triggerRunnerConfiguration = DefaultTriggerRunnerConfig
+  protected def triggerRunnerConfiguration: TriggerRunnerConfig = DefaultTriggerRunnerConfig
 
-  protected def ledgerClient(
-      maxInboundMessageSize: Int = RunnerConfig.DefaultMaxInboundMessageSize,
-      config: Option[LedgerClientConfiguration] = None,
-  )(implicit ec: ExecutionContext): Future[LedgerClient] = {
-    val effectiveConfig = config.getOrElse(ledgerClientConfiguration)
-    for {
-      client <- LedgerClient
-        .singleHost(
-          "localhost",
-          serverPort.value,
-          effectiveConfig,
-          ledgerClientChannelConfiguration.copy(maxInboundMessageSize = maxInboundMessageSize),
-        )
-    } yield client
-  }
-
-  override protected def darFile =
-    Try(BazelRunfiles.requiredResource("triggers/tests/acs.dar"))
-      .getOrElse(BazelRunfiles.requiredResource("triggers/tests/acs-1.dev.dar"))
-
-  protected val dar = DarDecoder.assertReadArchiveFromFile(darFile)
-  protected val compiledPackages =
-    PureCompiledPackages.assertBuild(dar.all.toMap, speedy.Compiler.Config.Dev)
+  protected val CompiledDar(packageId, compiledPackages) =
+    readDar(darFile.merge, speedy.Compiler.Config.Dev)
 
   protected def getRunner(
       client: LedgerClient,
       name: QualifiedName,
-      party: String,
-      readAs: Set[String] = Set.empty,
+      party: Party,
+      readAs: Set[Party] = Set.empty,
   ): Runner = {
     val triggerId = Identifier(packageId, name)
 
     Trigger.newTriggerLogContext(
       triggerId,
-      Party(party),
-      Party.subst(readAs),
+      party,
+      readAs,
       "test-trigger",
-      ApplicationId("test-trigger-app"),
+      applicationId,
     ) { implicit triggerContext: TriggerLogContext =>
       val trigger = Trigger.fromIdentifier(compiledPackages, triggerId).toOption.get
 
@@ -119,11 +104,11 @@ trait AbstractTriggerTest
         trigger,
         triggerRunnerConfiguration,
         client,
-        config.participants(ParticipantId).apiServer.timeProviderType,
+        timeProviderType,
         applicationId,
         TriggerParties(
-          actAs = Party(party),
-          readAs = Party.subst(readAs),
+          actAs = party,
+          readAs = readAs,
         ),
       )
     }
@@ -140,14 +125,20 @@ object AbstractTriggerTest {
       config: SValue,
   )
 
-  def allocateParty(client: LedgerClient)(implicit ec: ExecutionContext): Future[String] =
-    client.partyManagementClient.allocateParty(None, None).map(_.party)
+  def allocateParty(
+      client: LedgerClient,
+      hint: Option[String] = None,
+      displayName: Option[String] = None,
+  )(implicit ec: ExecutionContext): Future[Party] =
+    client.partyManagementClient
+      .allocateParty(hint, displayName)
+      .map(details => Party(details.party: String))
 
-  def queryACS(client: LedgerClient, party: String)(implicit
+  def queryACS(client: LedgerClient, party: Party)(implicit
       ec: ExecutionContext,
       materializer: Materializer,
   ): Future[Map[LedgerApi.Identifier, Seq[LedgerApi.Record]]] = {
-    val filter = TransactionFilter(List((party, Filters.defaultInstance)).toMap)
+    val filter = TransactionFilter(List((party.unwrap, Filters.defaultInstance)).toMap)
     val contractsF: Future[Seq[CreatedEvent]] = client.activeContractSetClient
       .getActiveContracts(filter, verbose = true)
       .runWith(Sink.seq)
@@ -162,7 +153,7 @@ object AbstractTriggerTest {
     )
   }
 
-  def create(client: LedgerClient, party: String, cmd: CreateCommand)(implicit
+  def create(client: LedgerClient, party: Party, cmd: CreateCommand)(implicit
       ec: ExecutionContext,
       applicationId: ApplicationId,
   ): Future[String] = {
@@ -170,10 +161,10 @@ object AbstractTriggerTest {
     val request = SubmitAndWaitRequest(
       Some(
         Commands(
-          party = party,
+          party = party.unwrap,
           commands = commands,
           ledgerId = client.ledgerId.unwrap,
-          applicationId = ApplicationId.unwrap(applicationId),
+          applicationId = applicationId.unwrap,
           commandId = UUID.randomUUID.toString,
         )
       )
@@ -183,9 +174,29 @@ object AbstractTriggerTest {
     } yield response.getTransaction.events.head.getCreated.contractId
   }
 
+  def create(
+      client: LedgerClient,
+      party: Party,
+      commands: Seq[CreateCommand],
+      elements: Int = 50,
+      per: FiniteDuration = 1.second,
+  )(implicit
+      ec: ExecutionContext,
+      materializer: Materializer,
+      applicationId: ApplicationId,
+  ): Future[Unit] = {
+    Source(commands)
+      .mapAsync(8) { cmd =>
+        create(client, party, cmd)
+      }
+      .throttle(elements, per)
+      .run()
+      .map(_ => ())
+  }
+
   def exercise(
       client: LedgerClient,
-      party: String,
+      party: Party,
       templateId: LedgerApi.Identifier,
       contractId: String,
       choice: String,
@@ -204,10 +215,10 @@ object AbstractTriggerTest {
     val request = SubmitAndWaitRequest(
       Some(
         Commands(
-          party = party,
+          party = party.unwrap,
           commands = commands,
           ledgerId = client.ledgerId.unwrap,
-          applicationId = ApplicationId.unwrap(applicationId),
+          applicationId = applicationId.unwrap,
           commandId = UUID.randomUUID.toString,
         )
       )
@@ -219,7 +230,7 @@ object AbstractTriggerTest {
 
   def archive(
       client: LedgerClient,
-      party: String,
+      party: Party,
       templateId: LedgerApi.Identifier,
       contractId: String,
   )(implicit ec: ExecutionContext, applicationId: ApplicationId): Future[Unit] = {
