--- triggers/tests/src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/AbstractTriggerTestWithCanton.scala
@@ -1,237 +1,0 @@
-// Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.lf
-package engine
-package trigger
-package test
-
-import java.util.UUID
-import akka.stream.scaladsl.{Sink, Source}
-import com.daml.bazeltools.BazelRunfiles
-import com.daml.ledger.api.refinements.ApiTypes.Party
-import com.daml.ledger.api.v1.command_service.SubmitAndWaitRequest
-import com.daml.ledger.api.v1.commands.{Command, CreateCommand, ExerciseCommand, _}
-import com.daml.ledger.api.v1.event.CreatedEvent
-import com.daml.ledger.api.v1.transaction_filter.{Filters, TransactionFilter}
-import com.daml.ledger.api.v1.{value => LedgerApi}
-import com.daml.ledger.client.LedgerClient
-import com.daml.ledger.client.configuration.{
-  CommandClientConfiguration,
-  LedgerClientChannelConfiguration,
-  LedgerClientConfiguration,
-  LedgerIdRequirement,
-}
-import com.daml.lf.data.Ref._
-import com.daml.lf.engine.trigger.TriggerRunnerConfig.DefaultTriggerRunnerConfig
-import com.daml.lf.integrationtest.CantonFixture
-import com.daml.lf.speedy.SValue
-import com.daml.lf.speedy.SValue._
-import org.scalatest._
-import scalaz.syntax.tag._
-
-import scala.concurrent.duration._
-import scala.concurrent.{ExecutionContext, Future}
-import scala.util.{Try, Success, Failure}
-
-// TODO: once test migration work has completed, rename this trait to AbstractTriggerTest
-trait AbstractTriggerTestWithCanton extends CantonFixture {
-  self: Suite =>
-
-  import CantonFixture._
-
-  private[this] lazy val darFile =
-    Try(BazelRunfiles.requiredResource("triggers/tests/acs.dar").toPath) match {
-      case Success(value) => Right(value)
-      case Failure(_) => Left(BazelRunfiles.requiredResource("triggers/tests/acs-1.dev.dar").toPath)
-    }
-
-  final override protected lazy val darFiles = List(darFile.merge)
-  final override protected lazy val devMode = darFile.isLeft
-
-  protected def toHighLevelResult(s: SValue) = s match {
-    case SRecord(_, _, values) if values.size == 6 =>
-      AbstractTriggerTestWithCanton.HighLevelResult(
-        values.get(0),
-        values.get(1),
-        values.get(2),
-        values.get(3),
-        values.get(4),
-        values.get(5),
-      )
-    case _ => throw new IllegalArgumentException(s"Expected record with 6 fields but got $s")
-  }
-
-  protected def ledgerClientConfiguration: LedgerClientConfiguration =
-    LedgerClientConfiguration(
-      applicationId = applicationId.unwrap,
-      ledgerIdRequirement = LedgerIdRequirement.none,
-      commandClient = CommandClientConfiguration.default,
-      token = None,
-    )
-
-  protected def ledgerClientChannelConfiguration: LedgerClientChannelConfiguration =
-    LedgerClientChannelConfiguration.InsecureDefaults
-
-  protected def triggerRunnerConfiguration: TriggerRunnerConfig = DefaultTriggerRunnerConfig
-
-  protected val CompiledDar(packageId, compiledPackages) =
-    readDar(darFile.merge, speedy.Compiler.Config.Dev)
-
-  protected def getRunner(
-      client: LedgerClient,
-      name: QualifiedName,
-      party: Party,
-      readAs: Set[Party] = Set.empty,
-  ): Runner = {
-    val triggerId = Identifier(packageId, name)
-
-    Trigger.newTriggerLogContext(
-      triggerId,
-      party,
-      readAs,
-      "test-trigger",
-      applicationId,
-    ) { implicit triggerContext: TriggerLogContext =>
-      val trigger = Trigger.fromIdentifier(compiledPackages, triggerId).toOption.get
-
-      Runner(
-        compiledPackages,
-        trigger,
-        triggerRunnerConfiguration,
-        client,
-        timeProviderType,
-        applicationId,
-        TriggerParties(
-          actAs = party,
-          readAs = readAs,
-        ),
-      )
-    }
-  }
-
-  def allocateParty(
-      client: LedgerClient,
-      hint: Option[String] = None,
-      displayName: Option[String] = None,
-  )(implicit ec: ExecutionContext): Future[Party] =
-    client.partyManagementClient
-      .allocateParty(hint, displayName)
-      .map(details => Party(details.party: String))
-
-  protected def create(client: LedgerClient, party: Party, cmd: CreateCommand)(implicit
-      ec: ExecutionContext
-  ): Future[String] = {
-    val commands = Seq(Command().withCreate(cmd))
-    val request = SubmitAndWaitRequest(
-      Some(
-        Commands(
-          party = party.unwrap,
-          commands = commands,
-          ledgerId = client.ledgerId.unwrap,
-          applicationId = applicationId.unwrap,
-          commandId = UUID.randomUUID.toString,
-        )
-      )
-    )
-    for {
-      response <- client.commandServiceClient.submitAndWaitForTransaction(request)
-    } yield response.getTransaction.events.head.getCreated.contractId
-  }
-
-  protected def create(
-      client: LedgerClient,
-      party: Party,
-      commands: Seq[CreateCommand],
-      elements: Int = 50,
-      per: FiniteDuration = 1.second,
-  )(implicit ec: ExecutionContext): Future[Unit] = {
-    Source(commands)
-      .mapAsync(8) { cmd =>
-        create(client, party, cmd)
-      }
-      .throttle(elements, per)
-      .run()
-      .map(_ => ())
-  }
-
-  protected def exercise(
-      client: LedgerClient,
-      party: Party,
-      templateId: LedgerApi.Identifier,
-      contractId: String,
-      choice: String,
-      choiceArgument: LedgerApi.Value,
-  )(implicit ec: ExecutionContext): Future[Unit] = {
-    val commands = Seq(
-      Command().withExercise(
-        ExerciseCommand(
-          templateId = Some(templateId),
-          contractId = contractId,
-          choice = choice,
-          choiceArgument = Some(choiceArgument),
-        )
-      )
-    )
-    val request = SubmitAndWaitRequest(
-      Some(
-        Commands(
-          party = party.unwrap,
-          commands = commands,
-          ledgerId = client.ledgerId.unwrap,
-          applicationId = applicationId.unwrap,
-          commandId = UUID.randomUUID.toString,
-        )
-      )
-    )
-    for {
-      _ <- client.commandServiceClient.submitAndWaitForTransaction(request)
-    } yield ()
-  }
-
-  protected def archive(
-      client: LedgerClient,
-      party: Party,
-      templateId: LedgerApi.Identifier,
-      contractId: String,
-  )(implicit ec: ExecutionContext): Future[Unit] = {
-    exercise(
-      client,
-      party,
-      templateId,
-      contractId,
-      "Archive",
-      LedgerApi.Value().withRecord(LedgerApi.Record()),
-    )
-  }
-
-  protected def queryACS(client: LedgerClient, party: Party)(implicit
-      ec: ExecutionContext
-  ): Future[Map[LedgerApi.Identifier, Seq[LedgerApi.Record]]] = {
-    val filter = TransactionFilter(List((party.unwrap, Filters.defaultInstance)).toMap)
-    val contractsF: Future[Seq[CreatedEvent]] = client.activeContractSetClient
-      .getActiveContracts(filter, verbose = true)
-      .runWith(Sink.seq)
-      .map(_.flatMap(x => x.activeContracts))
-    contractsF.map(contracts =>
-      contracts
-        .map(created => (created.getTemplateId, created.getCreateArguments))
-        .groupBy(_._1)
-        .view
-        .mapValues(cs => cs.map(_._2))
-        .toMap
-    )
-  }
-
-}
-
-object AbstractTriggerTestWithCanton {
-  final case class HighLevelResult(
-      acs: SValue,
-      party: SValue,
-      readAs: SValue,
-      state: SValue,
-      commandsInFlight: SValue,
-      config: SValue,
-  )
-}
